requirements.txt


flask
numpy
librosa
soundfile
scipy
tensorflow-cpu
